refactor(accounts): Use a single file for all domain errors
