--- src/accounts/domain/errors/institutional_id_in_use.go
@@ -1,18 +1,0 @@
-package errors
-
-import (
-	"fmt"
-	"net/http"
-)
-
-type InstitutionalIdAlreadyInUseError struct {
-	InstitutionalId string
-}
-
-func (err InstitutionalIdAlreadyInUseError) Error() string {
-	return fmt.Sprintf("Institutional ID %s is already in use", err.InstitutionalId)
-}
-
-func (err InstitutionalIdAlreadyInUseError) StatusCode() int {
-	return http.StatusConflict
-}
